fix: refactor component
