+++ src/app/components/whiteboards-index/whiteboards-index.css
@@ -11,7 +11,6 @@
   width: 100%;
   max-width: 10000px;
   max-height: 7000px;
-  /* background-color: #e9e9e9; */
   background-color: white;
 }
 
@@ -24,32 +23,6 @@
   font-size: 20px;
 }
 
-.whiteboard-item {
-  /* max-height: 60px; */
-  transition: max-height 0.7s ease;
-}
-
-.whiteboard-item:hover {
-  /* max-height: 200px; */
-}
-
 .whiteboards_create_button:hover {
   filter: brightness(1.1);
-}
-
-.button-option {
-  background-color: white;
-  margin-left: 10px;
-  padding: 5px;
-  cursor: pointer;
-  border-radius: 10px;
-  border: none;
-  box-shadow: rgba(0, 0, 0, 0.25) 0px 1px 2px;
-  /* background: rgb(255, 255, 255);
-  background: linear-gradient(
-    180deg,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(211, 211, 211, 1) 80%,
-    rgba(211, 211, 211, 1) 100%
-  ); */
 }
